Refactored add button in to the same location as list items
