+++ src/client/components/CompanyLocalHeader/LocalHeaderCompanyLists.jsx
@@ -1,46 +1,73 @@
 import React from 'react'
 import styled from 'styled-components'
-import { BLUE, DARK_BLUE_LEGACY } from '../../../client/utils/colours'
+import {
+  DARK_BLUE_LEGACY,
+  GREY_3,
+  GREY_3_LEGACY,
+  LIGHT_BLUE_25,
+} from '../../../client/utils/colours'
 import { FONT_SIZE } from '@govuk-react/constants'
 import { connect } from 'react-redux'
 import { state2props } from '../CompanyLists/state'
 import Task from '../Task'
 import { COMPANY_LISTS__LISTS_LOADED } from '../../actions'
 
-const StyledCompanyListItemButton = styled('button')`
+const StyledCompanyListButton = styled('button')`
   display: inline-table;
   padding: 8px;
-  background-color: ${BLUE};
   border: none;
-  border-bottom: 5px solid ${DARK_BLUE_LEGACY};
   vertical-align: middle;
   cursor: pointer;
+  margin-right: 10px;
   span {
     pointer-events: none;
-    display: table-cell;
+    display: inline-block;
     font-size: ${FONT_SIZE.SIZE_16};
   }
 `
 
-export const LocalHeaderCompanyLists = ({ lists }) => (
-  <Task.Status
-    name="Company lists"
-    id="local-header"
-    progressMessage="Loading my companies lists"
-    startOnRender={{ onSuccessDispatch: COMPANY_LISTS__LISTS_LOADED }}
-  >
-    {() =>
-      lists && (
-        <>
-          {Object.keys(lists).map((key) => (
-            <StyledCompanyListItemButton key={key}>
-              {lists[key].name}
-            </StyledCompanyListItemButton>
-          ))}
-        </>
-      )
-    }
-  </Task.Status>
-)
+const StyledCompanyListItemButton = styled(StyledCompanyListButton)`
+  colour: ${DARK_BLUE_LEGACY};
+  background-color: ${LIGHT_BLUE_25};
+  border-bottom: 5px solid ${DARK_BLUE_LEGACY};
+`
+
+const StyledAddButton = styled(StyledCompanyListButton)`
+  background-color: ${GREY_3};
+  border-bottom: 5px solid ${GREY_3_LEGACY};
+`
+
+export const LocalHeaderCompanyLists = ({ lists, company, returnUrl }) => {
+  const queryString = returnUrl ? `${returnUrl}` : `/companies/${company.id}`
+  const handleClick = () => {
+    window.location.href = `/companies/${company.id}/lists/add-remove?returnUrl=${queryString}`
+  }
+  return (
+    <Task.Status
+      name="Company lists"
+      id="local-header"
+      progressMessage="Loading my companies lists"
+      startOnRender={{ onSuccessDispatch: COMPANY_LISTS__LISTS_LOADED }}
+    >
+      {() =>
+        lists && (
+          <>
+            {Object.keys(lists).map((list) => (
+              <StyledCompanyListItemButton key={list}>
+                {lists[list].name} <span>x</span>
+              </StyledCompanyListItemButton>
+            ))}
+            <StyledAddButton
+              data-test={'add-to-list-button'}
+              onClick={handleClick}
+            >
+              <span>+</span> Add to list
+            </StyledAddButton>
+          </>
+        )
+      }
+    </Task.Status>
+  )
+}
 
 export default connect(state2props)(LocalHeaderCompanyLists)
